feat(core,api): inject platform capabilities via ModuleInitContext.services
Add PlatformServices to @appstrate/core/module — a typed surface listing every
capability modules may need from the platform (orchestrator, pubsub, models,
packages, runs, realtime, cross-module events, logger). Populate it in
apps/api/src/lib/modules/registry.ts::buildModuleInitContext so modules
receive a pre-wired reference at init().

Also extract ContainerOrchestrator, PubSub, and related workload types into
@appstrate/core/platform-types. apps/api re-exports those type-only symbols
from their original paths — zero call-site changes.

Why: external modules (loaded via MODULES=<npm-pkg>) can now consume platform
runtime capabilities via ctx.services instead of reaching into apps/api
internals. Core stays framework-agnostic — the concrete service signatures
live in apps/api; core exposes only the structural contract.

Built-in modules (oidc, webhooks) are untouched — they continue to use
relative imports since they ship in-tree. The injection pattern is strictly
additive for them.
