+++ apps/api/src/infra/pubsub/interface.ts
@@ -5,9 +5,4 @@
  * Implementations: Redis (multi-instance) and local EventEmitter (single-instance).
  */
 
-export interface PubSub {
-  publish(channel: string, message: string): Promise<void>;
-  subscribe(channel: string, handler: (message: string) => void): Promise<void>;
-  unsubscribe(channel: string): Promise<void>;
-  shutdown(): Promise<void>;
-}
+export type { PubSub } from "@appstrate/core/platform-types";
