+++ apps/api/src/services/orchestrator/interface.ts
@@ -1,62 +1,3 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import type {
-  WorkloadHandle,
-  WorkloadSpec,
-  IsolationBoundary,
-  SidecarConfig,
-  CleanupReport,
-  StopResult,
-} from "./types.ts";
-
-export interface ContainerOrchestrator {
-  /** Init one-shot: pool init, platform detection, etc. */
-  initialize(): Promise<void>;
-
-  /** Graceful shutdown: drain pool, release resources. */
-  shutdown(): Promise<void>;
-
-  /** Clean up orphaned workloads/networks from a previous crash. */
-  cleanupOrphans(): Promise<CleanupReport>;
-
-  /** Ensure images are locally available (pull if missing/outdated). No-op when not applicable. */
-  ensureImages(images: string[]): Promise<void>;
-
-  /** Create an isolated environment for a run. Docker: bridge network. K8s: namespace. */
-  createIsolationBoundary(runId: string): Promise<IsolationBoundary>;
-
-  /** Remove an isolated environment. Idempotent. */
-  removeIsolationBoundary(boundary: IsolationBoundary): Promise<void>;
-
-  /**
-   * Create a ready-to-use sidecar (from pool or fresh).
-   * Internally handles: pool acquisition, fresh creation, platform network,
-   * ExtraHosts, port mapping, health check.
-   */
-  createSidecar(
-    runId: string,
-    boundary: IsolationBoundary,
-    config: SidecarConfig,
-  ): Promise<WorkloadHandle>;
-
-  /** Create a workload (agent). Does NOT start it — file injection included in the spec. */
-  createWorkload(spec: WorkloadSpec, boundary: IsolationBoundary): Promise<WorkloadHandle>;
-
-  /** Start a created workload. */
-  startWorkload(handle: WorkloadHandle): Promise<void>;
-
-  /** Stop a workload. Idempotent. */
-  stopWorkload(handle: WorkloadHandle, timeoutSeconds?: number): Promise<void>;
-
-  /** Remove a workload. Idempotent. */
-  removeWorkload(handle: WorkloadHandle): Promise<void>;
-
-  /** Wait for a workload to finish. Returns the exit code. */
-  waitForExit(handle: WorkloadHandle): Promise<number>;
-
-  /** Stream logs from a running workload. Format-agnostic (text line by line). */
-  streamLogs(handle: WorkloadHandle, signal?: AbortSignal): AsyncGenerator<string>;
-
-  /** Stop ALL workloads for a run by ID. For cancel. */
-  stopByRunId(runId: string, timeoutSeconds?: number): Promise<StopResult>;
-}
+export type { ContainerOrchestrator } from "@appstrate/core/platform-types";
